fggfgf

+++ Assets/System/Prefabs Extra/XDhp.cs
@@ -52,7 +52,7 @@ public class XDhp : ComponentBehavior
             fillMed.fillAmount = ratio;
         }
         
-        hp.text = mon.GetPar("health").ToString();
+        hp.text = ((int)mon.GetPar("health")).ToString();
         
         fill.gameObject.SetActive(ratio < 1 || !nofull);
         hp.gameObject.SetActive(!notext);
